+++ packages/caip/src/adapters/coingecko/utils.ts
@@ -10,11 +10,19 @@ import {
   CHAIN_NAMESPACE,
   CHAIN_REFERENCE,
   cosmosChainId,
+  dogeChainId,
   ethAssetId,
   ethChainId,
+  ltcChainId,
   osmosisChainId
 } from '../../constants'
-import { makeBtcData, makeCosmosHubData, makeOsmosisData } from '../../utils'
+import {
+  bitcoinAssetMap,
+  cosmosAssetMap,
+  dogecoinAssetMap,
+  litecoinAssetMap,
+  osmosisAssetMap
+} from '../../utils'
 
 export type CoingeckoCoin = {
   id: string
@@ -24,16 +32,6 @@ export type CoingeckoCoin = {
 }
 
 type AssetMap = Record<ChainId, Record<AssetId, string>>
-
-export const writeFiles = async (data: AssetMap) => {
-  await Promise.all(
-    Object.entries(data).map(async ([chainId, assets]) => {
-      const path = `./src/adapters/coingecko/generated/${chainId}/adapter.json`.replace(':', '_')
-      await fs.promises.writeFile(path, JSON.stringify(assets))
-    })
-  )
-  console.info('Generated CoinGecko AssetId adapter data.')
-}
 
 export const fetchData = async (URL: string) => (await axios.get<CoingeckoCoin[]>(URL)).data
 
@@ -78,8 +76,20 @@ export const parseData = (coins: CoingeckoCoin[]): AssetMap => {
 
   return {
     ...assetMap,
-    [btcChainId]: makeBtcData(),
-    [cosmosChainId]: makeCosmosHubData(),
-    [osmosisChainId]: makeOsmosisData()
+    [btcChainId]: bitcoinAssetMap,
+    [dogeChainId]: dogecoinAssetMap,
+    [ltcChainId]: litecoinAssetMap,
+    [cosmosChainId]: cosmosAssetMap,
+    [osmosisChainId]: osmosisAssetMap
   }
+}
+
+export const writeFiles = async (data: AssetMap) => {
+  await Promise.all(
+    Object.entries(data).map(async ([chainId, assets]) => {
+      const path = `./src/adapters/coingecko/generated/${chainId}/adapter.json`.replace(':', '_')
+      await fs.promises.writeFile(path, JSON.stringify(assets))
+    })
+  )
+  console.info('Generated CoinGecko AssetId adapter data.')
 }